# Conflicts:
#	GeneticAlgorithm_TSP/src/sample/Calculations.java

+++ GeneticAlgorithm_TSP/src/sample/Route.java
@@ -3,7 +3,7 @@ package sample;
 public class Route {
     private int distance;
     private int [] path;
-    private double fitness = 0;
+    private int fitness = 0;
 
     Route(int tot_distance, int[] path) {
         this.distance = tot_distance;
@@ -12,8 +12,8 @@ public class Route {
 
     public int get_distance() {return distance;}
     public int[] getPath() {return path;}
-    public double getFitness() {return fitness;}
+    public int getFitness() {return fitness;}
 
-    public void setFitness(double fitness) { this.fitness = fitness; }
+    public void setFitness(int fitness) { this.fitness = fitness; }
 
 }
